refactor: move radius_graph to datasets/graph_utils/spherenet_graph_utils, clean docstring
